Moves blooms package into index package
