+++ src/apps/chifra/pkg/index/bloom.go
@@ -1,4 +1,4 @@
-package bloom
+package index
 
 // Copyright 2021 The TrueBlocks Authors. All rights reserved.
 // Use of this source code is governed by a license that can
