--- src/apps/chifra/pkg/index/bloom/doc.go
@@ -1,2 +1,0 @@
-// Package bloom implementes a bloom filter
-package bloom
