+++ src/apps/chifra/pkg/index/bloom_ismember.go
@@ -1,4 +1,4 @@
-package bloom
+package index
 
 import (
 	"encoding/binary"
