+++ src/apps/chifra/pkg/index/bloom_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a license that can
 // be found in the LICENSE file.
 
-package bloom
+package index
 
 import (
 	"fmt"
